add FunctionJobCreationTaskStatus enum to handle task status convertion
