+++ services/api-server/src/simcore_service_api_server/_service_function_jobs_task_client.py
@@ -1,8 +1,9 @@
 # pylint: disable=too-many-instance-attributes
 import contextlib
+import logging
 from dataclasses import dataclass
-from typing import Final
 
+from celery_library.errors import TaskNotFoundError
 from common_library.exclude import as_dict_exclude_none
 from models_library.functions import (
     FunctionClass,
@@ -32,6 +33,9 @@ from models_library.rpc_pagination import PageLimitInt
 from models_library.users import UserID
 from servicelib.celery.models import TaskMetadata, TasksQueue, TaskUUID
 from servicelib.celery.task_manager import TaskManager
+from simcore_service_api_server.models.schemas.functions import (
+    FunctionJobCreationTaskStatus,
+)
 from sqlalchemy.ext.asyncio import AsyncEngine
 
 from ._service_function_jobs import FunctionJobService
@@ -43,15 +47,13 @@ from .exceptions.function_errors import (
 )
 from .models.api_resources import JobLinks
 from .models.domain.celery_models import ApiWorkerTaskFilter
+from .models.schemas.functions import FunctionJobCreationTaskStatus
 from .models.schemas.jobs import JobInputs, JobPricingSpecification
 from .services_http.webserver import AuthSession
 from .services_rpc.storage import StorageService
 from .services_rpc.wb_api_server import WbApiRpcClient
 
-_JOB_CREATION_TASK_STATUS_PREFIX: Final[str] = "JOB_CREATION_TASK_STATUS_"
-_JOB_CREATION_TASK_NOT_YET_SCHEDULED_STATUS: Final[str] = (
-    f"{_JOB_CREATION_TASK_STATUS_PREFIX}NOT_YET_SCHEDULED"
-)
+_logger = logging.getLogger(__name__)
 
 
 def join_inputs(
@@ -75,15 +77,18 @@ async def _celery_task_status(
     product_name: ProductName,
 ) -> str:
     if job_creation_task_id is None:
-        return _JOB_CREATION_TASK_NOT_YET_SCHEDULED_STATUS
+        return FunctionJobCreationTaskStatus.NOT_YET_SCHEDULED
     task_filter = ApiWorkerTaskFilter(
         user_id=user_id,
         product_name=product_name,
     )
-    task_status = await task_manager.get_task_status(
-        task_uuid=TaskUUID(job_creation_task_id), task_filter=task_filter
-    )
-    return f"{_JOB_CREATION_TASK_STATUS_PREFIX}{task_status.task_state}"
+    try:
+        task_status = await task_manager.get_task_status(
+            task_uuid=TaskUUID(job_creation_task_id), task_filter=task_filter
+        )
+        return FunctionJobCreationTaskStatus[task_status.task_state].value
+    except TaskNotFoundError as e:
+        return FunctionJobCreationTaskStatus.ERROR
 
 
 @dataclass(frozen=True, kw_only=True)
