+++ services/api-server/src/simcore_service_api_server/api/dependencies/models_schemas_function_filters.py
@@ -8,7 +8,7 @@ from models_library.functions import (
     FunctionJobCollectionsListFilters,
     FunctionJobID,
 )
-from simcore_service_api_server.models.schemas.functions_filters import (
+from simcore_service_api_server.models.schemas.functions import (
     FunctionJobsListFilters,
 )
