+++ services/api-server/src/simcore_service_api_server/models/schemas/functions.py
@@ -1,0 +1,54 @@
+from enum import StrEnum
+from typing import Annotated, Final
+
+from models_library.functions import FunctionID, FunctionJobCollectionID, FunctionJobID
+from pydantic import BaseModel, ConfigDict, Field
+from servicelib.celery.models import TaskState
+
+RUN_JOB_TASK_STATUS_PREFIX: Final[str] = "RUN_JOB_TASK_STATUS_"
+
+
+class FunctionJobsListFilters(BaseModel):
+    """Filters for listing function jobs"""
+
+    function_id: Annotated[
+        FunctionID | None,
+        Field(
+            description="Filter by function ID pattern",
+        ),
+    ] = None
+
+    function_job_ids: Annotated[
+        list[FunctionJobID] | None,
+        Field(
+            description="Filter by function job IDs",
+        ),
+    ] = None
+
+    function_job_collection_id: Annotated[
+        FunctionJobCollectionID | None,
+        Field(
+            description="Filter by function job collection ID",
+        ),
+    ] = None
+
+    model_config = ConfigDict(
+        extra="ignore",
+    )
+
+
+class FunctionJobCreationTaskStatus(StrEnum):
+    PENDING = f"{RUN_JOB_TASK_STATUS_PREFIX}PENDING"
+    STARTED = f"{RUN_JOB_TASK_STATUS_PREFIX}STARTED"
+    RETRY = f"{RUN_JOB_TASK_STATUS_PREFIX}RETRY"
+    SUCCESS = f"{RUN_JOB_TASK_STATUS_PREFIX}SUCCESS"
+    FAILURE = f"{RUN_JOB_TASK_STATUS_PREFIX}FAILURE"
+    NOT_YET_SCHEDULED = (
+        "Function job creation task not yet scheduled"  # api-server custom status
+    )
+    ERROR = "Function job creation failed"  # api-server custom status
+
+
+assert {elm._name_ for elm in TaskState}.union({"NOT_YET_SCHEDULED", "ERROR"}) == {
+    elm._name_ for elm in FunctionJobCreationTaskStatus
+}
